removing some lint warnings
git-svn-id: https://svn.testnett.uninett.no/radsecproxy/branches/release-1.3@440 e88ac4ed-0b26-0410-9574-a7f39faa03bf

+++ radsecproxy.h
@@ -181,7 +181,7 @@ struct rewrite {
 struct protodefs {
     char *name;
     char *secretdefault;
-    uint8_t socktype;
+    int socktype;
     char *portdefault;
     uint8_t retrycountdefault;
     uint8_t retrycountmax;
